+++ NEWS.md
@@ -7,6 +7,7 @@ AppVeyor.
 * Added precomp function to avoid repeated computation.
 * mrs_data2mat function now collapses all dimensions to dynamics by default.
 * mrs_data objects now store the nucleus.
+* Added a reader for old Varian format (fid/procpar) data.
 
 # spant 1.7.0
 * Added write_mrs function which guesses the output format from the file